Added turret. Added GUI. Added bullets and minimap.
Game now has a turret that the missiles fire out of. The turret roughly
aims where the mouse is pointed. Also There is a GUI that keeps track of
Score, Lives, Health, and Missile. Helicopter can now shoot bullets to
pop balloons. Finally there is now a minimap keeping track of where the
player is in the map.